# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Rumah Sakit",
        "description": "API ini digunakan untuk mengelola data pasien, dokter, konsultasi, dan lainnya.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server API Lokal"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/rekam-medis": {
            "get": {
                "tags": [
                    "Rekam Medis"
                ],
                "summary": "Menampilkan semua rekam medis",
                "operationId": "d23d3249a6e0172185629e0425c0ca44",
                "responses": {
                    "200": {
                        "description": "Berhasil menampilkan semua rekam medis"
                    }
                }
            },
            "post": {
                "tags": [
                    "Rekam Medis"
                ],
                "summary": "Membuat data rekam medis baru",
                "operationId": "8d3b0bd074c34f9296f940f677b94169",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pasien_id",
                                    "tanggal_kunjungan",
                                    "diagnosis",
                                    "tindakan",
                                    "obat"
                                ],
                                "properties": {
                                    "pasien_id": {
                                        "type": "integer"
                                    },
                                    "tanggal_kunjungan": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "diagnosis": {
                                        "type": "string"
                                    },
                                    "tindakan": {
                                        "type": "string"
                                    },
                                    "obat": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rekam medis berhasil dibuat"
                    }
                }
            }
        },
        "/api/rekam-medis/{id}": {
            "get": {
                "tags": [
                    "Rekam Medis"
                ],
                "summary": "Menampilkan detail rekam medis",
                "operationId": "2c1b5e363c3cfd3d97ad0eda55e76886",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID rekam medis",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail rekam medis ditemukan"
                    },
                    "404": {
                        "description": "Rekam medis tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rekam Medis"
                ],
                "summary": "Memperbarui data rekam medis",
                "operationId": "1bb0b34bcf4b3bcfbe838fc277b24931",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID rekam medis",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pasien_id": {
                                        "type": "integer"
                                    },
                                    "tanggal_kunjungan": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "diagnosis": {
                                        "type": "string"
                                    },
                                    "tindakan": {
                                        "type": "string"
                                    },
                                    "obat": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data rekam medis berhasil diperbarui"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rekam Medis"
                ],
                "summary": "Menghapus data rekam medis",
                "operationId": "bceaaa368c3fbec6e785edf39ed673a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID rekam medis",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rekam medis berhasil dihapus"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Menampilkan semua user",
                "operationId": "ef3192335324c07faff67807a2e8a2f8",
                "responses": {
                    "200": {
                        "description": "Daftar semua user"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Membuat user baru",
                "operationId": "287bc3ec931e652c54bb2ddb777040bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Admin"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User berhasil dibuat"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Menampilkan detail user",
                "operationId": "430618624b90f3b94e90e86921470fdd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail user"
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Memperbarui data user",
                "operationId": "479f33f268cca6dba677c396b495b60d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User berhasil diperbarui"
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Menghapus user",
                "operationId": "fb7c6cd121c930b66eefc7569b0f0ab3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User berhasil dihapus"
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    }
                }
            }
        }
    },
...
}